+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.2] - 2026-01-16
+
+### Added
+
+- Custom comment user@host vs Keyname as comment
+- Include the type in the ssh Key comment (default: false) - syntax: `id_<type>_<name>` example: `id_ed25519_my_server_key`
+
 ## [0.0.1] - 2026-01-16
 
 ### Added